fix ut

+++ pkg/sql/colexec/group/group.go
@@ -161,7 +161,7 @@ func (ctr *container) generateAggStructures(ap *Argument) error {
 	var err error
 	i := 0
 	for i < len(ap.Aggs) {
-		if ctr.bat.Aggs[i], err = agg.NewAggWithConfig(int64(ap.Aggs[i].Op), ap.Aggs[i].Dist, []types.Type{*ctr.aggVecs[i].vec.GetType()}, ap.Aggs[i].Config); err != nil {
+		if ctr.bat.Aggs[i], err = agg.NewAggWithConfig(ap.Aggs[i].Op, ap.Aggs[i].Dist, []types.Type{*ctr.aggVecs[i].vec.GetType()}, ap.Aggs[i].Config); err != nil {
 			ctr.bat = nil
 			return err
 		}
